Moved `ParsedAddress` to subfolder
