+++ src/Address/Parser/AddressParser.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Address\Parser;
 
-use CultuurNet\UDB3\Address\ParsedAddress;
-
 interface AddressParser
 {
     public function parse(string $formattedAddress): ?ParsedAddress;
