+++ src/Address/Parser/GeopuntAddressParser.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Address\Parser;
 
-use CultuurNet\UDB3\Address\ParsedAddress;
 use CultuurNet\UDB3\Json;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Uri;
