+++ tests/Address/Parser/GeopuntAddressParserTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Address\Parser;
 
-use CultuurNet\UDB3\Address\ParsedAddress;
 use CultuurNet\UDB3\Json;
 use CultuurNet\UDB3\TestLogger;
 use GuzzleHttp\Psr7\Request;
