+++ tests/Organizer/ReadModel/RDF/OrganizerJsonToTurtleConverterTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Organizer\ReadModel\RDF;
 
 use CultuurNet\UDB3\Address\Parser\AddressParser;
-use CultuurNet\UDB3\Address\ParsedAddress;
+use CultuurNet\UDB3\Address\Parser\ParsedAddress;
 use CultuurNet\UDB3\Iri\CallableIriGenerator;
 use CultuurNet\UDB3\Model\Serializer\Organizer\OrganizerDenormalizer;
 use CultuurNet\UDB3\ReadModel\DocumentRepository;
